+++ example/src/components/icons/format_italic.js
@@ -1,0 +1,1 @@
+export default 'M10 4v3h2.21l-3.42 8H6v3h8v-3h-2.21l3.42-8H18V4z';