+++ example/src/pages/button-group/index.js
@@ -1,0 +1,3 @@
+import ButtonGroup from './ButtonGroup.svelte';
+
+export default ButtonGroup;