+++ rollup.config.js
@@ -1,5 +1,6 @@
 import svelte from 'rollup-plugin-svelte';
 import resolve from 'rollup-plugin-node-resolve';
+import { terser } from 'rollup-plugin-terser';
 import pkg from './package.json';
 
 const name = pkg.name
@@ -15,6 +16,7 @@ export default {
 	],
 	plugins: [
 		svelte(),
-		resolve()
+		resolve(),
+		terser()
 	]
 };