+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # svelte-ui changelog
 
+## 0.1.7
+
+- ToggleButton new component
+- ButtonGroup new component
+
 ## 0.1.5
 
 - Menu